Simplify 'if' description (#3173)
Simplify type precondition for 'if'

Fixes: #3152

Simplify description of if-statement structure

Refactor the grammar to match the prose.
This relies on some LALR mechanics, in the sense that "else" leads
to a parse state (item set) which can later accept either else_if_clause
or else_clause. That's fine in LALR(1) but would need modification for
an LL or recursive descent parser.

Move the introduction sentences of if-statement to the top of the
section. That's better for readability (and the other statement kinds
should be updated that way too)