Separate content/device internal states of GPUBuffer (#3300)
This change tries not to have behavior effects compared to what we have specced now, but since what we have now doesn't quite make sense, it picks just one of several possible interpretations of what we had written. #3271 is open about which one we actually pick. **I'm calling this editorial despite picking an interpretation it leaves an open issue to figure out which one to use.**

This is now independent of exposing the buffer state as an attribute (#3014). This change defines a number of content-side states from which that attribute would be defined.

The notable changes are:

- Server side now has just 2 states: available, unavailable.
- Client side has two state variables:
    - `[[mapping]]` which is either null, or all of the stuff for an actively-mapped buffer (not pending).
    - `[[map_requests]]` which tracks all map requests that haven't yet gotten a response from the device.
- At least in the mapping methods, client states are only accessed on the client, and server states are only accessed on the server.
- `unmap()` is rewritten to deinterlace the client and server side steps.
    - **`unmap()` doesn't reject promises. Instead it marks requests as cancelled so they get rejected when the map request completes.**