Don't throw exception when getBindGroupLayout index > maxBindGroups
This can be handled by returning an invalid GPUBindGroupLayout instead,
as we do if the index is past the end of the actual BGL list (just
below in the device timeline validation steps).

Since getBindGroupLayout returns a new JS wrapper every time, there's no
concern about that behavior in this case. I suspect that's why we had an
exception to begin with.