wgsl: Move * and & from lhs_expression to core_lhs_expression
This resolves an ambiguity on how to type-check things like:

    var m:mat2x2<f32>;
    let p = &m;
    *p[0] = m[1];

This grammar change pulls processing of the * and & from lhs_expression
into core_lhs_expression. So this example will parse as if:

   (*p)[0] = m[1];

This is necessary, unless we want to support #1580 (which propooses to
add apply array indexing and struct member selection on pointer values).

Note: This means reference expressions are parsed differently on the
left side of an assignment than they anywhere else.

Fixes: #3530