wgsl: describe partial references that still cover all memory locations
This can occur for single-member structures, and single-element arrays.
The locations in such a partial reference cover the whole variable, just like
a full reference does.  But the store type is the member type or element type (respectively).

And say that when analyzing uniformity of function-scope variables,
assignment through a partial reference is treated as if it's a full assignment.
This makes the analysis conservative.

This resolves #3605 by saying it does *not* relax the rule for
what a full reference can be.

Fixes: #3605