WGSL: allow bgra8unorm as texel format with extension bgra8unorm_storage
This patch intends to add `bgra8unorm` as a valid texel format when
the extension `bgra8unorm_storage` is enabled. `bgra8unorm_storage`
corresponds to the WebGPU extension `bgra8unorm-storage`.

Note that because in SPIRV there is no `bgra8` image format, the texel
format `bgra8unorm` should be translated into `rgba8` in SPIRV, and
in Vulkan backend an image with format rgba8unorm and component
swizzling (`r` to `b` and `b` to `r`) will be bound to the descriptor
set.