Add maxBindGroupsPlusVertexBuffers limit
Fixes https://github.com/gpuweb/gpuweb/issues/2749

On an implementation on top of Metal that chooses to use argument buffers,
bind groups and vertex buffers use the same metal resource (vertex buffer
slots). This limit allows the site to trade-off bind groups and vertex
buffers themselves, as long as the total sum of them don't exceed a limit.

This will be a bit tricky to implement on Metal if it's possible to unbind
vertex buffers or bind groups (see
https://github.com/gpuweb/gpuweb/issues/3787) but I think it's possible and
worth the tradeoff to not expose the complexity to the web, but to instead
just handle it in the browser.