grammar analyzer: Fix first and follow sets for degenerate rules (#3797)
Sometimes handwritten grammars have simple rules like:

    A -> B

or
    C -> D | ....

where B and D are symbols.

The First and Follow sets were not propagating across that kind of
construction.  Normally a rule is constructed canonically as a choice
over sequences. But in this case the rule is itself a single "Symbol"
or a choice where one of the alternatives is just a Symbol.

The First set of B should propagate to the First set of A.
The First set of D should propagate to the First set of C.

The Follow set of A should propagate to the Follow set of B.
The Follow set of C should propagate to the Follow set of D.

Also fix dumping a grammar to stdout.