Merge dfns in headings with the headings themselves
Bikeshed handles this pattern in accordance with the "Definitions data
model": https://speced.github.io/bikeshed/#dfn-contract

Note, previously each of these had two ids: one for the heading, one for the dfn. Some but not all of these are preserved:
- `#gpudevice` (was the dfn) instead of `#gpu-device` (was the heading)
- `#gpudevicedescriptor` (was the heading) instead of `#dictdef-gpudevicedescriptor` (was the dfn)
- `#gpuimagedatalayout` (did not exist before) instead of `#gpu-image-data-layout` (was the heading) or `#dictdef-gpuimagedatalayout` (was the dfn)
  - Same for `#gpuimagecopybuffer`, `#gpuimagecopytexture`, `#gpuimagecopytexturetagged`, `#gpuimagecopyexternalimage`
- `#gpuprogrammablestage` (did not exist before) instead of `#GPUProgrammableStage` (was the heading) or `#dictdef-gpuprogrammablestage` (was the dfn)
  - Same for `#gpucanvasalphamode`

We can preserve some of these historical URLs if we want. Should we?