Define "software extensions", and the "requires" directive
Define "extension" more generically, with two kinds:
 - enable-extensions: intended for hardware features, and must be
   requested via an enable directive.
 - software extensions: intended for features that could be feasibly
   supported anywhere.  They do not have to be requested.
   However, a "requires" directive documents the program's intended
   use of software extensions.

Extension names are context-dependent names, and are not identifiers,
and hence do not resolve to declarations.
No longer stop them from being spelled the same as keywords or reserved
words. (We're about to make f16 not a keyword).

Add "requires" as a keyword.

Reserve the word "require". Regenerate the reserved words list.

Fixes: #3149