Suppress unused dfns using noexport (#4001)
dneto0 taught me that we can suppress unused dfn warnings using noexport
instead of having this user-visible "temporary usages of non-exported
dfns" section at the bottom of the spec. I was just looking at getting
rid of that the other day and I wasn't sure how. This is much better!