Always connect the uniformity graph after a function call. (#4218)
* Always connect the uniformity graph after a function call.

For function calls, if the call was tagged as CallSiteRequiredToBeUniform.S
(any severity) then no edge was added to link the before (last CF_i) and
after (CF_after) the call. This was ok when the severity was always an
error, but if the severity is warning or info, the rules will miss a
collective operation requirement after the function call at a higher
severity. It should always be safe to connect the control flow before
and after. The disconnection was simply an optimization when diagnostics
were always errors.

* [editorial] Value constructors are functions now

* Update outdated statement since value constructors were made built-in
  functions