Specify copy rules for more external types
WebGL contexts use context.drawingBufferWidth and
context.drawingBufferHeight.

ImageBitmapRenderingContexts use their internal
"output bitmap"'s width and height (which are
opaque to the app) but the app itself
created added the ImageBitmap to the context
and so had it's chance to know the size.

HTMLImageElement uses naturalWidth and naturalHeight

ImageData uses width, height