Disallow requiring alignment limits >= 2**32 (#4358)
This text used to allow requesting alignment limits like 2**32 or 2**53.
This didn't do anything useful, but it adds implementation complexity,
because, per this spec, alignment limits are 32-bit integers. In order
to implement that you would need a special case to handle the casting of
the GPUSize64 (from the record<>) to a uint32_t internally.