Clarify the use of `@blend_src` in certain shader stages
It is not suitable to say `blend_src` "must only be used as an
output of the fragment shader stage" because apparantly we need to
use the struct variables with `blend_src` in many other places.

For example, we can first declare a temporary variable whose type
is a struct with `blend_src` and assign values into it before using
it in the return expression of the entry point.