Cross-reference from "domain" to "differences from IEEE754"
Restructure the paragraphs about domain into a definition and a bulleted
list.  This helps to scope the cross-reference mention of what WGSL
does when evaluating an operation outside its domain.

Also, link from the definition of "extended reals" to the Floating Point
Evaluation section. Do this because we already have a sentence about how
fp types approximate the extended reals.  So link to the whole fp eval
section so the reader can quickly jump to the rules.