Remove GPUShaderModuleDescriptor.sourceMap
TBD exactly how we will support source maps in the future, but we
realized this isn't the correct way.

If there are any applications out there setting this member, this is a
non-breaking change because browsers will just ignore the key.
(No browsers were implementing this anyway, so at best we were just
type-checking it.)

History: added in 645

Fixes 4808