Add API surface for feature levels
This provides a dictionary member that can be used for feature level
requests in the future. It currently does not accept any value; the only
thing it does is make sure that _if_ a feature level is requested, it
can be rejected. This is the appropriate behavior for a browser to have
before it adds support for any other feature levels.

It has type `any` so that it `requestAdapter()` can return `null`
(rather than rejecting) regardless of how we end up phrasing feature
levels in the future.

This is planned to be used for Compat mode, likely simply by accepting a
string value of `"compat"`.

Issue: 4656