Allow only strings in featureLevel
The initial landing of GPURequestAdapterOptions.featureLevel allowed
`any` because we don't know what the shape of future feature level
requests will be.

Brandon pointed out that we don't need to use `any` to achieve this.
Instead, we can add sibling members such that you request something
like:
`{ featureLevel: 'foo', featureLevelFooOptions: { /* ... */ } }`

There was a similar change in WebXR's history; discussion on that seems
to start roughly here:
https://github.com/immersive-web/webxr/issues/1205#issuecomment-868869651

(It is a DOMString rather than an enum so that we can return `null`
meaning "cannot fulfill adapter request" rather than rejecting. This
makes it consistent between older browsers, and newer browsers on older
hardware.)

Issue: 4656