Compat: update new limits proposal
See the proposal:

Effectively, in both compat and core, requesting limits for
`maxStorageBuffersInFragmentStage` and/or `maxStorageBuffersInVertexStage`
that are greater than `maxStorageBuffersPerShaderStage` is an error.

Similarly, in both compat and core, requesting limit for
`maxStorageTexturesInFragmentStage` and/or `maxStorageTexturesInVertexStage`
that are greater than `maxStorageTexturesPerShaderStage` is an error.

Assume the adapter supports 16 of each:

```js
requiredLimits: {
  maxStorageBuffersInFragmentStage: 3,
}
```

In compat this return a device with

```js
limits: {
  maxStorageBuffersInFragmentStage: 3,
  maxStorageBuffersPerShaderStage: 4,
}
```

In core it returns a device with

```js
limits: {
  maxStorageBuffersInFragmentStage: 4,
  maxStorageBuffersPerShaderStage: 4,
}
```

Another example:

```js
requiredLimits: {
  maxStorageBuffersInFragmentStage: 9,
  maxStorageBuffersPerShaderStage: 10,
}
```

In compat this return a device with

```js
limits: {
  maxStorageBuffersInFragmentStage: 9,
  maxStorageBuffersPerShaderStage: 10,
}
```

In core it returns a device with

```js
limits: {
  maxStorageBuffersInFragmentStage: 10,
  maxStorageBuffersPerShaderStage: 10,
}
```