Move Correspondence Reference from LD to UD, to fix CSS
`Status: LD` is "Living Document", which makes sense but has broken CSS
if used with `Group: webgpu`. If not used with `Group: webgpu` then the
license is CC0 (Public Domain) which is probably fine but not really
appropriate.

`Status: UD` is "Unofficial Proposal Draft", which is not really what I
would call this document, but it renders correctly :)

Any other status causes an error message saying something like: "You
used Status NOTE, but your Group (WEBGPU) is limited to the statuses
CG-DRAFT, CG-FINAL, or UD." (Probably LD isn't supposed to be used
either and that's why it's broken.)

Documentation: https://speced.github.io/bikeshed/#metadata