Adjust metadata and workflow to resume publication to /TR
New spec drafts need to be published as "Candidate Recommendation Drafts".
This status requires a deadline date to be specified. That deadline could
be set as a `W3C_BUILD_OVERRIDE` parameter in the workflow, but then it's
inherent to the specs, so it seems good to have it defined in the source
directly. (That information is not useful for the Editor's Draft in theory,
but also needed in practice since Bikeshed replaces all macro texts in the
boilerplate before it drops the bits that aren't needed).

The update also adds a link to the test suite in WebGPU and renames "Tests"
into "Test Suite" in WGSL (usual name across specs).