Define WGSL assertions as their own kind (#5054)
* Define assertions as their own kind rather than declarations or statements

This PR treats const_assert as its own kind of construct rather than as a declaration or statement. It introduces a new "assertions" section in the specification and defines const_assert as the only kind of assertion. The changes ensure that const_assert maintains consistent behavior whether used at module scope or within function scope.

* Adjust

* Update preview

* Relocate + subsume

* Update wgsl/index.bs

Co-authored-by: alan-baker <alanbaker@google.com>

* Update wgsl/index.bs

Co-authored-by: alan-baker <alanbaker@google.com>

---------

Co-authored-by: alan-baker <alanbaker@google.com>