Add "core-features-and-limits" to spec (#5147)
Add this to the main spec so that browsers can start
implementing/shipping it. This allows applications to start relying on
the presence of this feature name as soon as browsers ship it.

Currently since Compat is in the spec yet, this is always available on
adapters and always auto-enabled on devices.

Fixes #5127