Fix custom dark mode styles for Bikeshed switcher
Bikeshed added a light/dark mode theme switcher in a floating sidebar
(in the lower left). Rather than change the media query (because that's
not possible), it changes the class on the body element and
enables/disables the dark.css stylesheet.

Currently this switcher only works in TR (see speced/bikeshed#3092), but
working around that locally, this makes our styles follow the switcher
properly.

Issue: 5178