Fix custom dark mode styles for Bikeshed switcher (#5183)
Bikeshed added a light/dark mode theme switcher in a floating sidebar
(in the lower left). Rather than change the media query (because that's
not possible), it changes the class on the body element and
enables/disables the dark.css stylesheet.

Currently this switcher only works in TR (see speced/bikeshed#3092), but
working around that locally, this makes our styles follow the switcher
properly. (Unfortunately since the switcher doesn't work on github.io,
this makes it so that IF you use it, it messes up these particular
styles. But that'll go away when it's fixed.)

Issue:
https://github.com/gpuweb/gpuweb/discussions/5178#discussioncomment-12987193