[immediates proposal] Rename to immediateSize and maxImmediateSize (#5173)
Rename the pipeline layout member to be simpler, and rename the limit to
match the usual limit naming pattern.

~Other limits all start with `max` (or `min`). Also, this is the
`max` `immediateDataRangeByteSize` so
`maxImmediateDataRangeByteSize`
is nice because it contains the full name of the thing it's the max of.~

I realized while adding this to `webgpu.h` that we needed to add the
limit, too. We hadn't scrutinized the naming of it.

Issue: #75