Add [AllowShared] to setBindGroup (#5202)
This was always intended to be here, we just forgot it when we were
adding SharedArrayBuffer support elsewhere (which went through many
iterations due to lack of support in WebIDL).

This is not technically an editorial change but we definitely intended
it, so IMO we should go ahead and land it and I'll put it as an FYI on
the WG agenda.

Fixes #5201