Define featureLevels; introduce Compat Mode CSS; give an example (#5198)
Give definitions for the "core" and "compatibility" featureLevels, as
well as "Core-defaulting" and "Compatibility-defaulting" Adapters
(language largely taken from the compat proposal).

Define CSS for "Compatibility Mode" exceptions: a cyan-ish box, similar
to the "Valid Usage" styling.
As an example, add restriction 2 ("Color blending state may not differ
between color attachments in a `GPUFragmentState`.") from the compat
proposal in that style.

Co-authored-by: Kai Ninomiya <kainino@chromium.org>
Co-authored-by: Kai Ninomiya <kainino1@gmail.com>