Fix template list discovery algorithm: missing continue
At the end of the first 'if' clause, where '<' is matched,
insert a 'Start the next iteration of the loop.'
Otherwise execution reaches the 'advance token' at the bottom
of the loop.

For example, in parsing:  X < Y >
The '>' would be consumed too early, and template list discovery
gets the wrong answer.

Fixes: #5246