[Compat] Fixes for Compat initialization and more precise prose (#5443)
This started with addressing François's editorial comment to define
"Compatibility Mode", but I determined that most of the references to
"Compatibility Mode" should actually be more descriptive and point more
precisely to relevant definitions, and that the definitions of the
feature level strings should be more comprehensive.

Also removed definitions of "Core-defaulting" and
"Compatibility-defaulting" in favor of direct links (but kept those
phrases in the prose as they're useful for understanding).

Additionally I noticed that were missing text to make
`"core-features-and-limits"` increase the limits, and text to explicitly
say we set the default limits depending on feature level. Fixed those.

Issue:
https://github.com/gpuweb/gpuweb/pull/5402#pullrequestreview-3410836743