Tweak requestAdapter step ordering, feature level definitions (again)
Following other styles, validation should come first, then other stuff.
Also tweak some wording to make clear what is implementation-defined,
and that the [=adapter=] is always a new object (necessarily, because
the expiry state is on it).

Make the feature level definitions more formal because they are depended
on by the requestAdapter steps. Make the "new device" steps defer to
those definitions instead of duplicating them.