uniformity: relax rules when loop bodies always break or return (#5419)
When a loop body always breaks or returns, control does not reach the
loop's continuing block. Adjust the uniformity graph accordingly.

Rework the uniformity analysis for loop statements:
* organize according to syntatic form: without continuing block, then
with continuing block.
* split rows based on behaviours of s1 instead of the whole loop. That
made the partition much clearer to me.
* everywhere, use the the fact that statement behaviour is a subset of
{Next,Continue,Break,Return}
* Split rows (at the end) based on whether s1's behaviour intersected
with {Next,Continue}, or not. If so, then you get the reverse-backedge
CF' -> CF1 (or CF2).
* In one case split a row based on whether s1's behaviour includes
Return, or not.

Adjust variable value analysis:
- Vin(s1) gets an edge from Vout(s2) only if behaviour of s1 includes
either Next or Continue
- Vin(s2) gets an edge from Vout(s1) only if behaviour of s1 includes
Next
- Vin(next) gets an edge from Vout(s2) only if behaviour of s1 includes
Break

Issue: #5364